sqrt ll -> long long

+++ source/math_other/sqrtll.cpp
@@ -1,8 +1,8 @@
 // floor(sqrt(n))
-ll sqrtll(ll n) { 
-    ll le = 0, ri = 3'000'000'001;
+long long sqrtll(long long n) { 
+    long long le = 0, ri = 3'000'000'001;
     while(ri - le > 1) {
-        ll mid = (le + ri) / 2;
+        long long mid = (le + ri) / 2;
         if(mid * mid <= n) le = mid;
         else ri = mid;
     }
